✨feat: JPA Entity 재배치 및 엔티티 간 참조 구현 (#9)
