+++ src/main/java/kr/modernworld/modernworldv2/admin/infrastructure/persistence/entity/BanJPAEntity.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.admin.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -25,7 +25,7 @@ public class BanJPAEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
+  private Long no;
 
   @Size(max = 300)
   @NotNull
