+++ src/main/java/kr/modernworld/modernworldv2/admin/infrastructure/persistence/entity/CharacterJPAEntity.java
@@ -1,11 +1,19 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.admin.infrastructure.persistence.entity;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import java.util.ArrayList;
+import java.util.List;
+import kr.modernworld.modernworldv2.user.infrastructure.persistence.entity.CharacterLockerJPAEntity;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
@@ -17,8 +25,9 @@ import org.hibernate.annotations.ColumnDefault;
 public class CharacterJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
+  private Long no;
 
   @Size(max = 15)
   @NotNull
@@ -43,5 +52,8 @@ public class CharacterJPAEntity {
   @ColumnDefault("'0'")
   @Column(name = "price", columnDefinition = "int UNSIGNED not null")
   private Long price;
+
+  @OneToMany(mappedBy = "character", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
+  private List<CharacterLockerJPAEntity> characterLockers = new ArrayList<>();
 
 }
