--- src/main/java/kr/modernworld/modernworldv2/global/persistence/enitity/UserAchievementJPAEntity.java
@@ -1,58 +1,0 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Index;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import jakarta.validation.constraints.NotNull;
-import java.time.Instant;
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "user_achievement", schema = "modernworld", indexes = {
-    @Index(name = "user_no", columnList = "user_no"),
-    @Index(name = "achievement_no", columnList = "achievement_no")
-}, uniqueConstraints = {
-    @UniqueConstraint(name = "user_achievement_user_no_achievement_no_key", columnNames = {
-        "user_no", "achievement_no"})
-})
-public class UserAchievementJPAEntity {
-
-  @Id
-  @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
-
-  @NotNull
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "user_no", nullable = false)
-  private UserJPAEntity userNo;
-
-  @NotNull
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "achievement_no", nullable = false)
-  private AchievementJPAEntity achievementNo;
-
-  @NotNull
-  @ColumnDefault("0")
-  @Column(name = "status", nullable = false)
-  private Boolean status = false;
-
-  @NotNull
-  @ColumnDefault("CURRENT_TIMESTAMP")
-  @Column(name = "created_at", nullable = false)
-  private Instant createdAt;
-
-}
