+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/AchievementLevel.java
@@ -1,0 +1,7 @@
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
+
+enum AchievementLevel {
+  one,
+  two,
+  three
+}
