+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/CharacterLockerJPAEntity.java
@@ -1,8 +1,10 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
@@ -10,7 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 import jakarta.validation.constraints.NotNull;
-import java.time.Instant;
+import kr.modernworld.modernworldv2.admin.infrastructure.persistence.entity.CharacterJPAEntity;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
@@ -20,34 +22,31 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "inventory", schema = "modernworld", indexes = {
-    @Index(name = "user_no", columnList = "user_no"),
-    @Index(name = "item_no", columnList = "item_no")
+@Table(name = "character_locker", schema = "modernworld", indexes = {
+    @Index(name = "character_no", columnList = "character_no"),
+    @Index(name = "user_no", columnList = "user_no")
 }, uniqueConstraints = {
-    @UniqueConstraint(name = "inventory_user_no_item_no_key", columnNames = {"user_no", "item_no"})
+    @UniqueConstraint(name = "character_locker_user_no_character_no_key", columnNames = {"user_no",
+        "character_no"})
 })
-public class InventoryJPAEntity {
+public class CharacterLockerJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
+  private Long no;
+
+  @NotNull
+  @ManyToOne(fetch = FetchType.LAZY, optional = false)
+  @OnDelete(action = OnDeleteAction.CASCADE)
+  @JoinColumn(name = "character_no", nullable = false)
+  private CharacterJPAEntity character;
 
   @NotNull
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
   @JoinColumn(name = "user_no", nullable = false)
-  private UserJPAEntity userNo;
-
-  @NotNull
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "item_no", nullable = false)
-  private ItemJPAEntity itemNo;
-
-  @NotNull
-  @ColumnDefault("CURRENT_TIMESTAMP")
-  @Column(name = "created_at", nullable = false)
-  private Instant createdAt;
+  private UserJPAEntity user;
 
   @NotNull
   @ColumnDefault("0")
