+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/CommentJPAEntity.java
@@ -1,16 +1,22 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
@@ -20,26 +26,26 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "reply", schema = "modernworld", indexes = {
-    @Index(name = "comment_no", columnList = "comment_no"),
-    @Index(name = "user_no", columnList = "user_no")
+@Table(name = "comment", schema = "modernworld", indexes = {
+    @Index(name = "receiver_no", columnList = "receiver_no"),
+    @Index(name = "sender_no", columnList = "sender_no")
 })
-public class ReplyJPAEntity {
+public class CommentJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
-
-  @NotNull
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "comment_no", nullable = false)
-  private CommentJPAEntity commentNo;
+  private Long no;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @OnDelete(action = OnDeleteAction.SET_NULL)
-  @JoinColumn(name = "user_no")
-  private UserJPAEntity userNo;
+  @JoinColumn(name = "receiver_no")
+  private UserJPAEntity receiver;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @OnDelete(action = OnDeleteAction.SET_NULL)
+  @JoinColumn(name = "sender_no")
+  private UserJPAEntity sender;
 
   @Size(max = 300)
   @NotNull
@@ -53,5 +59,8 @@ public class ReplyJPAEntity {
 
   @Column(name = "deleted_at")
   private Instant deletedAt;
+
+  @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
+  private List<ReplyJPAEntity> replies = new ArrayList<>();
 
 }
