+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/GameResult.java
@@ -1,0 +1,7 @@
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
+
+enum GameResult {
+  win,
+  lose,
+  draw
+}
