+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/InventoryJPAEntity.java
@@ -1,15 +1,19 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
 import jakarta.validation.constraints.NotNull;
 import java.time.Instant;
+import kr.modernworld.modernworldv2.admin.infrastructure.persistence.entity.ItemJPAEntity;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
@@ -19,36 +23,39 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "neighbor", schema = "modernworld", indexes = {
-    @Index(name = "sender_no", columnList = "sender_no"),
-    @Index(name = "receiver_no", columnList = "receiver_no")
+@Table(name = "inventory", schema = "modernworld", indexes = {
+    @Index(name = "user_no", columnList = "user_no"),
+    @Index(name = "item_no", columnList = "item_no")
+}, uniqueConstraints = {
+    @UniqueConstraint(name = "inventory_user_no_item_no_key", columnNames = {"user_no", "item_no"})
 })
-public class NeighborJPAEntity {
+public class InventoryJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
+  private Long no;
 
   @NotNull
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "sender_no", nullable = false)
-  private UserJPAEntity senderNo;
+  @JoinColumn(name = "user_no", nullable = false)
+  private UserJPAEntity user;
 
   @NotNull
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "receiver_no", nullable = false)
-  private UserJPAEntity receiverNo;
-
-  @NotNull
-  @ColumnDefault("0")
-  @Column(name = "status", nullable = false)
-  private Boolean status = false;
+  @JoinColumn(name = "item_no", nullable = false)
+  private ItemJPAEntity item;
 
   @NotNull
   @ColumnDefault("CURRENT_TIMESTAMP")
   @Column(name = "created_at", nullable = false)
   private Instant createdAt;
+
+  @NotNull
+  @ColumnDefault("0")
+  @Column(name = "status", nullable = false)
+  private Boolean status = false;
 
 }
