+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/LikeJPAEntity.java
@@ -1,8 +1,10 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
@@ -27,19 +29,20 @@ import org.hibernate.annotations.OnDeleteAction;
 public class LikeJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
+  private Long no;
 
   @NotNull
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
   @JoinColumn(name = "receiver_no", nullable = false)
-  private UserJPAEntity receiverNo;
+  private UserJPAEntity receiver;
 
   @NotNull
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
   @JoinColumn(name = "sender_no", nullable = false)
-  private UserJPAEntity senderNo;
+  private UserJPAEntity sender;
 
 }
