+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/PostJPAEntity.java
@@ -1,15 +1,17 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import java.time.Instant;
 import lombok.Getter;
 import lombok.Setter;
@@ -20,31 +22,31 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "present", schema = "modernworld", indexes = {
-    @Index(name = "item_no", columnList = "item_no"),
+@Table(name = "post", schema = "modernworld", indexes = {
     @Index(name = "sender_no", columnList = "sender_no"),
     @Index(name = "receiver_no", columnList = "receiver_no")
 })
-public class PresentJPAEntity {
+public class PostJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
-
-  @NotNull
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @JoinColumn(name = "item_no", nullable = false)
-  private ItemJPAEntity itemNo;
+  private Long no;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @OnDelete(action = OnDeleteAction.SET_NULL)
   @JoinColumn(name = "sender_no")
-  private UserJPAEntity senderNo;
+  private UserJPAEntity sender;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @OnDelete(action = OnDeleteAction.SET_NULL)
   @JoinColumn(name = "receiver_no")
-  private UserJPAEntity receiverNo;
+  private UserJPAEntity receiver;
+
+  @Size(max = 150)
+  @NotNull
+  @Column(name = "content", nullable = false, length = 150)
+  private String content;
 
   @NotNull
   @ColumnDefault("CURRENT_TIMESTAMP")
@@ -52,10 +54,9 @@ public class PresentJPAEntity {
   private Instant createdAt;
 
   @NotNull
-  @ColumnDefault("'unread'")
-  @Lob
-  @Column(name = "status", nullable = false)
-  private String status;
+  @ColumnDefault("0")
+  @Column(name = "`check`", nullable = false)
+  private Boolean check = false;
 
   @NotNull
   @ColumnDefault("0")
