+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/PresentJPAEntity.java
@@ -1,16 +1,20 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import java.time.Instant;
+import kr.modernworld.modernworldv2.admin.infrastructure.persistence.entity.ItemJPAEntity;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
@@ -20,30 +24,32 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "post", schema = "modernworld", indexes = {
+@Table(name = "present", schema = "modernworld", indexes = {
+    @Index(name = "item_no", columnList = "item_no"),
     @Index(name = "sender_no", columnList = "sender_no"),
     @Index(name = "receiver_no", columnList = "receiver_no")
 })
-public class PostJPAEntity {
+public class PresentJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
+  private Long no;
+
+  @NotNull
+  @ManyToOne(fetch = FetchType.LAZY, optional = false)
+  @JoinColumn(name = "item_no", nullable = false)
+  private ItemJPAEntity item;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @OnDelete(action = OnDeleteAction.SET_NULL)
   @JoinColumn(name = "sender_no")
-  private UserJPAEntity senderNo;
+  private UserJPAEntity sender;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @OnDelete(action = OnDeleteAction.SET_NULL)
   @JoinColumn(name = "receiver_no")
-  private UserJPAEntity receiverNo;
-
-  @Size(max = 150)
-  @NotNull
-  @Column(name = "content", nullable = false, length = 150)
-  private String content;
+  private UserJPAEntity receiver;
 
   @NotNull
   @ColumnDefault("CURRENT_TIMESTAMP")
@@ -51,9 +57,10 @@ public class PostJPAEntity {
   private Instant createdAt;
 
   @NotNull
-  @ColumnDefault("0")
-  @Column(name = "`check`", nullable = false)
-  private Boolean check = false;
+  @ColumnDefault("'unread'")
+  @Enumerated(EnumType.STRING)
+  @Column(name = "status", nullable = false)
+  private PresentStatus status;
 
   @NotNull
   @ColumnDefault("0")
