+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/PresentStatus.java
@@ -1,0 +1,8 @@
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
+
+enum PresentStatus {
+  unread,
+  read,
+  accept,
+  reject
+}
