+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/ReplyJPAEntity.java
@@ -1,0 +1,60 @@
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import java.time.Instant;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "reply", schema = "modernworld", indexes = {
+    @Index(name = "comment_no", columnList = "comment_no"),
+    @Index(name = "user_no", columnList = "user_no")
+})
+public class ReplyJPAEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "no", columnDefinition = "int UNSIGNED not null")
+  private Long no;
+
+  @NotNull
+  @ManyToOne(fetch = FetchType.LAZY, optional = false)
+  @OnDelete(action = OnDeleteAction.CASCADE)
+  @JoinColumn(name = "comment_no", nullable = false)
+  private CommentJPAEntity comment;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @OnDelete(action = OnDeleteAction.SET_NULL)
+  @JoinColumn(name = "user_no")
+  private UserJPAEntity user;
+
+  @Size(max = 300)
+  @NotNull
+  @Column(name = "content", nullable = false, length = 300)
+  private String content;
+
+  @NotNull
+  @ColumnDefault("CURRENT_TIMESTAMP")
+  @Column(name = "created_at", nullable = false)
+  private Instant createdAt;
+
+  @Column(name = "deleted_at")
+  private Instant deletedAt;
+
+}
