+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/ReportCategory.java
@@ -1,0 +1,15 @@
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
+
+enum ReportCategory {
+  spam,
+  harmfulContent,
+  scamImpersonation,
+  copyrightInfringement,
+  explicitContent,
+  abusiveBehavior,
+  misinformation,
+  duplicateContent,
+  hateSpeech,
+  technicalIssue,
+  other
+}
