+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/persistence/entity/UserAchievementJPAEntity.java
@@ -1,8 +1,10 @@
-package kr.modernworld.modernworldv2.global.persistence.enitity;
+package kr.modernworld.modernworldv2.user.infrastructure.persistence.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
@@ -10,6 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 import jakarta.validation.constraints.NotNull;
+import java.time.Instant;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
@@ -19,34 +22,40 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "character_locker", schema = "modernworld", indexes = {
-    @Index(name = "character_no", columnList = "character_no"),
-    @Index(name = "user_no", columnList = "user_no")
+@Table(name = "user_achievement", schema = "modernworld", indexes = {
+    @Index(name = "user_no", columnList = "user_no"),
+    @Index(name = "achievement_no", columnList = "achievement_no")
 }, uniqueConstraints = {
-    @UniqueConstraint(name = "character_locker_user_no_character_no_key", columnNames = {"user_no",
-        "character_no"})
+    @UniqueConstraint(name = "user_achievement_user_no_achievement_no_key", columnNames = {
+        "user_no", "achievement_no"})
 })
-public class CharacterLockerJPAEntity {
+public class UserAchievementJPAEntity {
 
   @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "no", columnDefinition = "int UNSIGNED not null")
-  private Long id;
-
-  @NotNull
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  @JoinColumn(name = "character_no", nullable = false)
-  private CharacterJPAEntity characterNo;
+  private Long no;
 
   @NotNull
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
   @JoinColumn(name = "user_no", nullable = false)
-  private UserJPAEntity userNo;
+  private UserJPAEntity user;
+
+  @NotNull
+  @ManyToOne(fetch = FetchType.LAZY, optional = false)
+  @OnDelete(action = OnDeleteAction.CASCADE)
+  @JoinColumn(name = "achievement_no", nullable = false)
+  private AchievementJPAEntity achievement;
 
   @NotNull
   @ColumnDefault("0")
   @Column(name = "status", nullable = false)
   private Boolean status = false;
+
+  @NotNull
+  @ColumnDefault("CURRENT_TIMESTAMP")
+  @Column(name = "created_at", nullable = false)
+  private Instant createdAt;
 
 }
